Trivial fix for style, and made the README and comment match up
Puppet-lint was complaining about "ensure found on line but it's not the
first attribute on line 109".  Trivial fix (swapped the ensure and the
name).

Also, the README didn't mention Debian or CentOS, but the comment block
in init.pp did.  I added both to the README.  I happen to be using this
with Debian unstable, but I haven't tested it extensively enough to say
"yep, this works".  It, however, works for me.